falconsai_nsfw/requirements.txt
torch
transformers
flask
pillow
gunicorn